UIImage - correct transform application
+Added new ui element : UIImage
+Added ImageProperty for UIImage
+Corrected the applyTransform handling, now take account of parent transform
+Added new functions to access new computed values in TransformProperty

+++ SFML-UIAddon/SFML-UIAddon-Dll/Code/RoundedRectangle.cpp
@@ -54,8 +54,8 @@ namespace sfui
 				point.y = m_size.y - finalRadius;
 				angle += 90;
 			}
-			point.x += cosf(angle * M_PI / 180) * finalRadius;
-			point.y += sinf(angle * M_PI / 180) * finalRadius;
+			point.x += cosf(angle * static_cast<float>(M_PI) / 180) * finalRadius;
+			point.y += sinf(angle * static_cast<float>(M_PI) / 180) * finalRadius;
 			return point;
 		}
 	}
